Move shortenFxName to fxChain.cpp.
